Move db models to their new home
