+++ event-handlers/src/main/java/no/sikt/nva/nvi/events/UpsertNviCandidateHandler.java
@@ -12,11 +12,11 @@ import java.net.URI;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import no.sikt.nva.nvi.common.model.business.DbCandidate;
-import no.sikt.nva.nvi.common.model.business.DbCreator;
-import no.sikt.nva.nvi.common.model.business.DbInstitutionPoints;
-import no.sikt.nva.nvi.common.model.business.DbLevel;
-import no.sikt.nva.nvi.common.model.business.DbPublicationDate;
+import no.sikt.nva.nvi.common.db.model.DbCandidate;
+import no.sikt.nva.nvi.common.db.model.DbCreator;
+import no.sikt.nva.nvi.common.db.model.DbInstitutionPoints;
+import no.sikt.nva.nvi.common.db.model.DbLevel;
+import no.sikt.nva.nvi.common.db.model.DbPublicationDate;
 import no.sikt.nva.nvi.common.service.NviService;
 import nva.commons.core.JacocoGenerated;
 import org.slf4j.Logger;
