+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/db/Candidate.java
@@ -2,8 +2,8 @@ package no.sikt.nva.nvi.common.db;
 
 import java.util.List;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.model.business.DbApprovalStatus;
-import no.sikt.nva.nvi.common.model.business.DbCandidate;
+import no.sikt.nva.nvi.common.db.model.DbApprovalStatus;
+import no.sikt.nva.nvi.common.db.model.DbCandidate;
 
 public record Candidate(
     UUID identifier,
