+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/db/model/DbReportingStatus.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.common.model.business;
+package no.sikt.nva.nvi.common.db.model;
 
 import java.net.URI;
 import java.time.Instant;
