+++ nvi-rest/src/main/java/no/sikt/nva/nvi/CandidateResponse.java
@@ -7,9 +7,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import no.sikt.nva.nvi.common.db.Candidate;
-import no.sikt.nva.nvi.common.model.business.DbApprovalStatus;
-import no.sikt.nva.nvi.common.model.business.DbCandidate;
-import no.sikt.nva.nvi.common.model.business.DbInstitutionPoints;
+import no.sikt.nva.nvi.common.db.model.DbApprovalStatus;
+import no.sikt.nva.nvi.common.db.model.DbCandidate;
+import no.sikt.nva.nvi.common.db.model.DbInstitutionPoints;
 import no.sikt.nva.nvi.fetch.ApprovalStatus;
 import no.sikt.nva.nvi.fetch.InstitutionPoints;
 import no.sikt.nva.nvi.fetch.Note;
