+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/CreateNviPeriodHandler.java
@@ -6,8 +6,8 @@ import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
 import no.sikt.nva.nvi.rest.utils.RequestUtil;
 import no.sikt.nva.nvi.rest.model.NviPeriodDto;
-import no.sikt.nva.nvi.common.model.business.DbNviPeriod;
-import no.sikt.nva.nvi.common.model.business.DbNviPeriod.Builder;
+import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.model.DbNviPeriod.Builder;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.utils.ExceptionMapper;
 import nva.commons.apigateway.AccessRight;
