+++ nvi-rest/src/main/java/no/sikt/nva/nvi/rest/UpdateNviPeriodHandler.java
@@ -4,8 +4,8 @@ import static no.sikt.nva.nvi.rest.utils.RequestUtil.getUsername;
 import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.net.HttpURLConnection;
-import no.sikt.nva.nvi.common.model.business.DbNviPeriod;
-import no.sikt.nva.nvi.common.model.business.DbNviPeriod.Builder;
+import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.model.DbNviPeriod.Builder;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.rest.utils.RequestUtil;
 import no.sikt.nva.nvi.rest.model.NviPeriodDto;
