+++ nvi-rest/src/test/java/no/sikt/nva/nvi/rest/UpdateNviPeriodHandlerTest.java
@@ -17,7 +17,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.util.NoSuchElementException;
-import no.sikt.nva.nvi.common.model.business.DbNviPeriod;
+import no.sikt.nva.nvi.common.db.model.DbNviPeriod;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.rest.model.NviPeriodDto;
 import no.unit.nva.commons.json.JsonUtils;
