+++ nvi-rest/src/test/java/no/sikt/nva/nvi/upsert/UpdateNviCandidateStatusHandlerDbTest.java
@@ -16,12 +16,12 @@ import java.util.List;
 import java.util.Map;
 import no.sikt.nva.nvi.CandidateResponse;
 import no.sikt.nva.nvi.common.db.NviCandidateRepository;
-import no.sikt.nva.nvi.common.model.business.DbApprovalStatus;
-import no.sikt.nva.nvi.common.model.business.DbCandidate;
-import no.sikt.nva.nvi.common.model.business.DbCreator;
-import no.sikt.nva.nvi.common.model.business.DbLevel;
-import no.sikt.nva.nvi.common.model.business.DbPublicationDate;
-import no.sikt.nva.nvi.common.model.business.DbStatus;
+import no.sikt.nva.nvi.common.db.model.DbApprovalStatus;
+import no.sikt.nva.nvi.common.db.model.DbCandidate;
+import no.sikt.nva.nvi.common.db.model.DbCreator;
+import no.sikt.nva.nvi.common.db.model.DbLevel;
+import no.sikt.nva.nvi.common.db.model.DbPublicationDate;
+import no.sikt.nva.nvi.common.db.model.DbStatus;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.rest.NviApprovalStatus;
 import no.sikt.nva.nvi.rest.NviStatusRequest;
