+++ nvi-rest/src/test/java/no/sikt/nva/nvi/upsert/UpdateNviCandidateStatusHandlerTest.java
@@ -24,10 +24,10 @@ import java.util.UUID;
 import java.util.stream.Stream;
 import no.sikt.nva.nvi.CandidateResponse;
 import no.sikt.nva.nvi.common.db.Candidate;
-import no.sikt.nva.nvi.common.model.business.DbApprovalStatus;
-import no.sikt.nva.nvi.common.model.business.DbCandidate;
-import no.sikt.nva.nvi.common.model.business.DbStatus;
-import no.sikt.nva.nvi.common.model.business.DbUsername;
+import no.sikt.nva.nvi.common.db.model.DbApprovalStatus;
+import no.sikt.nva.nvi.common.db.model.DbCandidate;
+import no.sikt.nva.nvi.common.db.model.DbStatus;
+import no.sikt.nva.nvi.common.db.model.DbUsername;
 import no.sikt.nva.nvi.common.service.NviService;
 import no.sikt.nva.nvi.fetch.ApprovalStatus;
 import no.sikt.nva.nvi.rest.NviApprovalStatus;
